{"schema":"decodex/commit/1","summary":"Modularize session selection interaction","authority":"manual"}
